requirements.txt
# Essential Steroid Components
customtkinter==5.2.2
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.2.0